+++ cella-lints/build.dart
@@ -1,22 +1,64 @@
 import 'dart:io';
 
+import 'package:cella_lints/cella_lints.dart';
 import 'package:path/path.dart';
 import 'package:yaml/yaml.dart';
 import 'package:yaml_edit/yaml_edit.dart';
 
 void main() {
+  final counter = TimeCounter.now();
+
+  // Get current repo and monorepo path.
   final current = Platform.script.toFilePath();
   final root = dirname(current);
   final monorepo = dirname(root);
 
-  // Copy license file into current child repo.
+  // Sync license and contributors.
   final license = 'license'.toUpperCase();
-  File(join(monorepo, license)).copySync(join(root, license));
-
-  // Sync contributors into current child repo.
-  final name = basename(root);
   final contributors = '${'contributors'.toUpperCase()}.yaml';
-  final path = join(monorepo, contributors);
+  copyLicense(root: root, monorepo: monorepo, licenseFilename: license);
+  syncContributors(
+    root: root,
+    monorepo: monorepo,
+    contributorsFilename: contributors,
+  );
+
+  // Auto generate .pubignore.
+  final content = [
+    contributors.toLowerCase(),
+    license.toLowerCase(),
+  ].map((item) => '!$item').join('\n');
+  final pubignorePath = join(root, '.pubignore');
+  File(pubignorePath).writeAsStringSync('$content\n');
+  stdout.writeln(
+    '${'pub ignore generated'.green.dim} '
+    '${pubignorePath.formatPath(style: (raw) => raw.blue)}',
+  );
+
+  counter.logDuration(message: 'done');
+}
+
+void copyLicense({
+  required String root,
+  required String monorepo,
+  required String licenseFilename,
+}) {
+  final pathFrom = join(monorepo, licenseFilename);
+  final pathTo = join(root, licenseFilename);
+  File(pathFrom).copySync(pathTo);
+  stdout.writeln(
+    '${'license copied'.green.dim} '
+    '${PathMove.from(pathFrom, pathTo).format((raw) => raw.magenta)}',
+  );
+}
+
+void syncContributors({
+  required String root,
+  required String monorepo,
+  required String contributorsFilename,
+}) {
+  final name = basename(root);
+  final path = join(monorepo, contributorsFilename);
   final raw = loadYaml(File(path).readAsStringSync());
   if (raw is! Map) throw Exception('invalid structure: $path');
   final handler = <String, dynamic>{};
@@ -31,12 +73,10 @@ void main() {
     }
   }
   final editor = YamlEditor('')..update([], handler);
-  File(join(root, contributors)).writeAsStringSync('$editor\n');
-
-  // Auto generate .pubignore.
-  final content = [
-    contributors.toLowerCase(),
-    license.toLowerCase(),
-  ].map((item) => '!$item').join('\n');
-  File(join(root, '.pubignore')).writeAsStringSync('$content\n');
+  final pathTo = join(root, contributorsFilename);
+  File(pathTo).writeAsStringSync('$editor\n');
+  stdout.writeln(
+    '${'contributors synced'.green.dim} '
+    '${PathMove.from(path, pathTo).format((raw) => raw.magenta)}',
+  );
 }