+++ cella-lints/lib/cella_lints.dart
@@ -1,2 +1,3 @@
+export 'src/format.dart';
 export 'src/option.dart';
 export 'src/sync.dart';